requirements.txt
asgiref==3.5.2
Django==4.1.2
django-environ==0.9.0
sqlparse==0.4.3
psycopg2==2.9.5
django-seed==0.3.1
djangorestframework==3.14.0
social-auth-app-django==5.0.0
Pillow==9.3.0
djangorestframework-simplejwt==5.2.2
django-password-validators==1.7.0
django-cors-headers==3.13.0
inflect==6.0.2
drf-yasg==1.21.4
coverage==7.1.0
pytest-django==4.5.2
mixer==7.2.2
